+++ HISTORY.md
@@ -1,5 +1,15 @@
 # history
 
+## 0.2.1
+
+* `NEW` add ok button
+* `FIXED` missing style
+* `CHANGED` onSelect will pass leafList
+
+## 0.2.0
+
+* `NEW` modal view
+
 ## 0.1.5
 
 * `FIXED` props.value rerender error